suggest unicode or closing LaTeX; end latex at a word
There are now two auto-suggestions for LaTeX: convert to Unicode (when
possible) and "LaTeX code", which just inserts whatever closing braces
and end delimiter are necessary.

The tokeniser to detect LaTeX code ends at a space followed by two or
more word characters, so that you can add LaTeX in the middle of a
previously written passage of text.